Separate gauges to different files
We want to keep different type of gauges in separate files to make code changes
easier in the future.
