+++ tools/metrics/gauges_cf.go
@@ -1,87 +1,13 @@
 package main
 
 import (
-	"crypto/tls"
 	"fmt"
 	"log"
-	"net"
 	"net/url"
-	"strings"
 	"time"
 
-	"code.cloudfoundry.org/lager"
-	"github.com/alphagov/paas-cf/tools/metrics/pingdumb"
-	"github.com/alphagov/paas-cf/tools/metrics/tlscheck"
 	cfclient "github.com/cloudfoundry-community/go-cfclient"
 )
-
-func TLSValidityGauge(logger lager.Logger, certChecker tlscheck.CertChecker, addr string, interval time.Duration) MetricReadCloser {
-	return NewMetricPoller(interval, func(w MetricWriter) error {
-		if !strings.Contains(addr, ":") {
-			addr += ":443"
-		}
-		host, _, err := net.SplitHostPort(addr)
-		if err != nil {
-			return err
-		}
-
-		daysUntilExpiry, err := certChecker.DaysUntilExpiry(addr, &tls.Config{})
-		if err != nil {
-			logger.Error("tls-certificates-validity-failure", err, lager.Data{
-				"addr":     addr,
-				"hostname": host,
-			})
-			return err
-		}
-
-		metric := Metric{
-			Kind:  Gauge,
-			Time:  time.Now(),
-			Name:  "tls.certificates.validity",
-			Value: daysUntilExpiry,
-			Tags: []string{
-				fmt.Sprintf("hostname:%s", host),
-			},
-		}
-		return w.WriteMetrics([]Metric{metric})
-	})
-}
-
-func CDNTLSValidityGauge(logger lager.Logger, certChecker tlscheck.CertChecker, cfs *CloudFrontService, interval time.Duration) MetricReadCloser {
-	return NewMetricPoller(interval, func(w MetricWriter) error {
-		customDomains, err := cfs.CustomDomains()
-		if err != nil {
-			logger.Error("cloudfront-list-distributions-failure", err, lager.Data{})
-			return err
-		}
-
-		metrics := []Metric{}
-		for _, customDomain := range customDomains {
-			daysUntilExpiry, err := certChecker.DaysUntilExpiry(
-				customDomain.CloudFrontDomain+":443",
-				&tls.Config{ServerName: customDomain.AliasDomain},
-			)
-			if err != nil {
-				logger.Error("cdn-tls-certificates-validity-failure", err, lager.Data{
-					"alias_domain":      customDomain.AliasDomain,
-					"cloudfront_domain": customDomain.CloudFrontDomain,
-				})
-				return err
-			}
-
-			metrics = append(metrics, Metric{
-				Kind:  Gauge,
-				Time:  time.Now(),
-				Name:  "cdn.tls.certificates.validity",
-				Value: daysUntilExpiry,
-				Tags: []string{
-					fmt.Sprintf("hostname:%s", customDomain.AliasDomain),
-				},
-			})
-		}
-		return w.WriteMetrics(metrics)
-	})
-}
 
 func QuotaGauge(c *Client, interval time.Duration) MetricReadCloser {
 	return NewMetricPoller(interval, func(w MetricWriter) error {
@@ -382,38 +308,6 @@ func OrgCountGauge(c *Client, interval time.Duration) MetricReadCloser {
 			})
 		}
 		return w.WriteMetrics(metrics)
-	})
-}
-
-func ELBNodeFailureCountGauge(logger lager.Logger, config pingdumb.ReportConfig, interval time.Duration) MetricReadCloser {
-	return NewMetricPoller(interval, func(w MetricWriter) error {
-		r, err := pingdumb.GetReport(config)
-		if err != nil {
-			return err
-		}
-		failures := r.Failures()
-		for _, failedCheck := range failures {
-			logger.Info("elb-node-failure", lager.Data{
-				"addr":  failedCheck.Addr,
-				"start": failedCheck.Start.Format(time.RFC3339Nano),
-				"stop":  failedCheck.Start.Format(time.RFC3339Nano),
-				"err":   failedCheck.Err().Error(),
-			})
-		}
-		return w.WriteMetrics([]Metric{
-			{
-				Kind:  Gauge,
-				Time:  time.Now(),
-				Name:  "aws.elb.unhealthy_node_count",
-				Value: float64(len(failures)),
-			},
-			{
-				Kind:  Gauge,
-				Time:  time.Now(),
-				Name:  "aws.elb.healthy_node_count",
-				Value: float64(len(r.Checks) - len(failures)),
-			},
-		})
 	})
 }
