+++ tools/metrics/gauges_tls_test.go
@@ -1,18 +1,14 @@
 package main_test
 
 import (
-	"context"
 	"crypto/tls"
 	"errors"
 	"fmt"
-	"net"
-	"net/http"
 	"time"
 
 	"code.cloudfoundry.org/lager"
 	. "github.com/alphagov/paas-cf/tools/metrics"
 	"github.com/alphagov/paas-cf/tools/metrics/fakes"
-	"github.com/alphagov/paas-cf/tools/metrics/pingdumb"
 	tlscheck_fakes "github.com/alphagov/paas-cf/tools/metrics/tlscheck/fakes"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/cloudfront"
@@ -22,19 +18,11 @@ import (
 	"github.com/onsi/gomega/gbytes"
 )
 
-var _ = Describe("Gauges", func() {
+var _ = Describe("TLS gauges", func() {
 
 	var (
-		resolvers []*net.Resolver
-		server    *http.Server
-		logger    lager.Logger
-		log       *gbytes.Buffer
-		records   = map[string][]string{
-			"gauges.test.": []string{
-				"127.0.0.1",
-				"0.0.0.0",
-			},
-		}
+		logger            lager.Logger
+		log               *gbytes.Buffer
 		tlsChecker        *tlscheck_fakes.FakeCertChecker
 		cloudFrontService *CloudFrontService
 		cloudFrontClient  *fakes.FakeCloudFrontAPI
@@ -82,29 +70,6 @@ var _ = Describe("Gauges", func() {
 		}
 	)
 
-	BeforeSuite(func() {
-		resolvers = []*net.Resolver{
-			&net.Resolver{
-				Dial: func(ctx context.Context, network, address string) (net.Conn, error) {
-					dialer := &net.Dialer{}
-					return dialer.DialContext(ctx, network, "127.0.0.1:8553")
-				},
-			},
-		}
-
-		go func() {
-			defer GinkgoRecover()
-			Expect(fakes.ListenAndServeDNS(":8553", records)).To(Succeed())
-		}()
-
-		server = fakes.ListenAndServeHTTP(":8580")
-	})
-
-	AfterSuite(func() {
-		fakes.ShutdownDNS()
-		server.Close()
-	})
-
 	BeforeEach(func() {
 		logger = lager.NewLogger("logger")
 		log = gbytes.NewBuffer()
@@ -112,46 +77,6 @@ var _ = Describe("Gauges", func() {
 		tlsChecker = &tlscheck_fakes.FakeCertChecker{}
 		cloudFrontClient = &fakes.FakeCloudFrontAPI{}
 		cloudFrontService = &CloudFrontService{Client: cloudFrontClient}
-	})
-
-	It("emits two metrics", func() {
-		config := pingdumb.ReportConfig{
-			Target:    "http://gauges.test:8580",
-			Resolvers: resolvers,
-			Timeout:   1 * time.Second,
-		}
-		gauge := ELBNodeFailureCountGauge(logger, config, 1*time.Second)
-		metric, err := gauge.ReadMetric()
-		Expect(err).NotTo(HaveOccurred())
-		Expect(metric.Name).To(Equal("aws.elb.unhealthy_node_count"))
-		Expect(metric.Value).To(Equal(float64(0)))
-
-		metric, err = gauge.ReadMetric()
-		Expect(err).NotTo(HaveOccurred())
-		Expect(metric.Name).To(Equal("aws.elb.healthy_node_count"))
-		numberOfAddrs := len(records["gauges.test."])
-		Expect(metric.Value).To(Equal(float64(numberOfAddrs)))
-	})
-
-	It("returns failure count > 0 and logs failures", func() {
-		config := pingdumb.ReportConfig{
-			Target:    "http://gauges.test:7878",
-			Resolvers: resolvers,
-			Timeout:   1 * time.Second,
-		}
-		gauge := ELBNodeFailureCountGauge(logger, config, 1*time.Second)
-		metric, err := gauge.ReadMetric()
-		Expect(err).NotTo(HaveOccurred())
-		Expect(metric.Name).To(Equal("aws.elb.unhealthy_node_count"))
-		numberOfAddrs := len(records["gauges.test."])
-		Expect(metric.Value).To(Equal(float64(numberOfAddrs)))
-
-		metric, err = gauge.ReadMetric()
-		Expect(err).NotTo(HaveOccurred())
-		Expect(metric.Name).To(Equal("aws.elb.healthy_node_count"))
-		Expect(metric.Value).To(Equal(float64(0)))
-
-		Expect(log).To(gbytes.Say(`"addr":\s*"\d+\.\d+.\d+\.\d+`))
 	})
 
 	Describe("TLS validity gauge", func() {
